+++ weblate/checks/base.py
@@ -56,12 +56,13 @@ class Check:
 
     def should_skip(self, unit):
         """Check whether we should skip processing this unit."""
+        all_flags = unit.all_flags
         # Is this check ignored
-        if self.ignore_string in unit.all_flags:
+        if self.ignore_string in all_flags or "ignore-all-checks" in all_flags:
             return True
 
         # Is this disabled by default
-        if self.default_disabled and self.enable_string not in unit.all_flags:
+        if self.default_disabled and self.enable_string not in all_flags:
             return True
 
         return False